+++ nodeconductor/cost_tracking/models.py
@@ -24,7 +24,6 @@ from nodeconductor.logging.loggers import LoggableMixin
 from nodeconductor.logging.models import AlertThresholdMixin
 from nodeconductor.structure import models as structure_models, SupportedServices
 
-
 logger = logging.getLogger(__name__)
 
 
@@ -123,7 +122,7 @@ class PriceEstimate(LoggableMixin, AlertThresholdMixin, core_models.UuidMixin, c
         if hasattr(self.scope, 'backend_id'):
             self.details['backend_id'] = self.scope.backend_id
         if self.is_resource_estimate():
-            self.details['service_name'] = self.scope.service_project_link.service.name
+            self.details['service_settings_name'] = self.scope.service_project_link.service.settings.name
             self.details['project_name'] = self.scope.service_project_link.project.name
         self.save(update_fields=['details'])
 
@@ -196,7 +195,7 @@ class PriceEstimate(LoggableMixin, AlertThresholdMixin, core_models.UuidMixin, c
 
     def get_scope_name(self):
         if self.scope:
-            if isinstance(self.scope, structure_models.ServiceProjectLink):
+            if isinstance(self.scope, (structure_models.ServiceProjectLink, structure_models.Service)):
                 # We need to display some meaningful name for SPL.
                 return str(self.scope)
             else: